* fix-histogram:
  histogram: change type from uint32_t to uint64_t

+++ source/Histogram.cpp
@@ -9,7 +9,7 @@ using namespace std;
 
 static uint16_t colorRange = 256;
 
-Histogram::Histogram(cv::Mat& image) : histogram(vector<uint32_t>(colorRange, 0))
+Histogram::Histogram(cv::Mat& image) : histogram(vector<uint64_t>(colorRange, 0))
 {
     // Initialize vector
     for (int64_t row = 0; row < image.rows; ++row)
@@ -20,8 +20,8 @@ Histogram::Histogram(cv::Mat& image) : histogram(vector<uint32_t>(colorRange, 0)
 color Histogram::getHighest()
 {
     color result = 0;
-    uint32_t counter = 0;
-    for (uint32_t i = 0; i < histogram.size(); ++i) {
+    uint64_t counter = 0;
+    for (uint64_t i = 0; i < histogram.size(); ++i) {
         if (histogram[i] > counter) {
             result = (color) i;
             counter = histogram[i];
@@ -34,7 +34,7 @@ color Histogram::getHighest()
 color Histogram::getLowest()
 {
     color result = 0;
-    uint32_t counter = numeric_limits<uint32_t>::max();
+    uint64_t counter = numeric_limits<uint64_t>::max();
     for (color i = 0; i < histogram.size(); ++i) {
         if (histogram[i] < counter) {
             result = i;
@@ -45,7 +45,7 @@ color Histogram::getLowest()
     return result;
 }
 
-uint32_t Histogram::at(color index) const
+uint64_t Histogram::at(color index) const
 {
     return histogram.at(index);
 }
